Implement parallel streaming download and construction of tar archive
This is still really raw.  Tons of warnings, and panics after the
downloads are done because we need to also wait for all of the in flight
processing of writes to the tar builder to land in object storage.

Still, for the first time I ran the command and the result was a local
tar archive with the data from multiple S3 objects.