Big step forward in testing infrastructure
Now there's a separate `ssstar-testing` crate with some test helpers
that will be useful for both integration and unit testing.  Also the
beginnings of integration testing for the create archive job, running
using a local minio instance.

Unfortunately the tests are not reliable because of a hyper bug that
seems to be caused by a tokio runtime somehow being shared across tests.
I can't figure out how that's happening.