Fix timing issue that was causing test unreliability
It's not reliable to hold a `Config` instance across multiple tests, but
that's just what the `ObjectStorageFactory` type was doing.  Had to drop
this singleton idea entirely, and take a `Config` argument every time an
object storage URL is parsed.  It's additional `clone`ing but it means
each test is completely isolated from the others, so if minio stops and
starts up on a new port between tests it won't suddenly break the world.