Add test to verify object upload parts events happen correctly
I was seeing some incorrect progress reporting behavior which I thought
was a bug in the library.  This change passed all existing tests, and I
determined the issue is actually that object upload progress reports are
generated for multiple objects in parallel, so the problem is the
progress rendering doesn't handle this correctly.