Fix bug with extract progress reporting
multiple objects can be uploading parts in parallel, but the progress
reporting code assumed only one object is ever uplaoded at a time (which
was true before the fix for #19 added more concurrency).  Now the
progress bar is redrawn to reflect the particulars of whatever object's
part was uploaded most recently.  This results in some flickering and
might be a bit confusing, but the alternative is to not display any
object-level progress at all (not ideal), or to render a separate
progress bar for each object being uploaded (could get messy)

This seems like a reasonable compromise for now