Change how extraction works for better concurrency (#40)
Change how extraction works for better concurrency

This is a substantial internal refactoring of the `extract` flow which
should dramatically increase extract performance in cases which
previously saw lower extract performance compared to create performance
for the same workload.

It completely rewrites the code that uploads from a tar archive to
object storage.  Now the max concurrency configured for the job will
always be used, even in cases of multi-part upload or a mix of
multi-part and single-part upload.  The code is more complicated, which
is why it wasn't initially implemented this way, but the result should
be substantially faster writes to object storage.

Does not close #19, but improves performance in some cases.  Actually resolving #19 will require more analysis and effort, however the changes here are valuable enough that I want to get them landed and released as we start to build S3 support into Elastio.