Add config operation to override user agent when talking to S3
We need to be able to send a custom user agent string to the S3 APIs as
part of our commercial product (Elastio) which uses `ssstar` internally.
There's an open issue to add this functionality to the AWS SDK for Rust
(awslabs/aws-sdk-rust#146), but it's very old and clearly has not been
implemented.

Through trial and error, and guided by the implementation in another
open source code base (awslabs/amazon-qldb-shell#196) I was able to
cobble something together that works.  It's not pretty, and it involves
mostly duplicating the internal implementation of the
`Client::with_confg` function from the S3 SDK, however it will have to
do for now until official support is added to the SDK.

Part of (but doesn't yet close) #94