Initial concept for enhanced ssstar API (#116)
* Update deps and reorganize

Update to the latest AWS SDK releases, dealing with the usual breakage
as a result.

Re-arrange dependencies and other workspace properties to avoid
duplication by using workspace-level properties.

Format TOML files with taplo consistent with the other Elastio projects.

Update to Rust toolchain 1.69

* Add new `storage` module and initial API design.

This new storage abstraction isn't implemented yet and isn't actually
part of the create or extract logic.  It's just the initial outline.

* Refactor a bit and make the tar builder report where the data goes

This is still a WIP but I was able to get code working reliably (and
with a test) that calculates where in the tar stream the data for each
object is stored.  That will be important in Elastio later when we need
to construct an index.

* Fix warnings and clippy lints

* Add byte offset to `tar_archive_object_written` progress reporter method

* Remove `tar_` prefix from progress methods.

Now that we have a new storage abstraction coming, the archive might not
actually be tar format so keeping the `tar_` prefix would lead to
confusion.

* Fix warnings and lints

* Uncomment some temporarily disabled test code

* Fix windows build

* Fix rustfmt (fuck!)

* Add object timestamp to progress callback

* Fix windows build harder

* Disable coverage since it's always broken

* Update `release.toml` for latest version of `cargo-release`