chore: Dependabot fixes and tidying up (#131)
I noticed that the Dependabot config is out of date and isn't working
right with the old format `rust-toolchain`.

I also took this opportunity to fix a broken link to releases in the
README

This configured dependabot to always add me as a reviewer. It also
switches from a weekly to a daily schedule.

I then got pulled into a yak-shaving exercise, ultimately updating several deps, updating to Rust 1.81, migrating to the `dtolnay/rust-toolchain` action, and updating `deny.toml` with some new licenses and some suppression of dupes.