Opt out of new impl trait lifetime behavior in one problematic place
In 2021 edition, `impl Trait` in return position was assumed not to use any type params or lifetimes, but now in 2024 edition it is permitted to use by any lifetimes or type params in the function signature.
Maybe that makes sense as a default, but there was one place where the old behavior was assumed, and broke under the new behavior.

See https://blog.rust-lang.org/2024/09/05/impl-trait-capture-rules.html#impl-traits-can-include-a-use-bound-to-specify-precisely-which-generic-types-and-lifetimes-they-use