Migrate to new cargo deny TOML format
I didn't want to do this as part of the Rust 2024 edition migration, but
it's so out of date now that it's breaking the build.  I could expend
some energy to modify the build to use a deliberately ancient version of
the tool, or I can use that to migrate to the new version.  Hopefully I
made the right decision.