Switch from cargo udeps to cargo-machete
This is another yak shave that isn't directly related to the update to
Rust 2024 edition but it getting in the way of landing it.  The GHA
build used an ancient version of cargo-udeps that worked on an ancient
version of nightly.  The rest of the company's projects have long since
moved on from this.  So I'm switching to cargo-machete, which found some
unused deps that I also removed.